Send packet responses back to GDB immediately
Previously the tcp_write() calls used by MRI to send packet responses
back to GDB would just be placed in a queue to see if more data would be
queued up later to be sent in the same TCP/IP packet. This nagle time is
quite high and caused terrible performance for the GDB debugging
experience. With this commit, the code now detects when blocks large
enough to be an empty GDB packet or larger are passed into tcp_write()
and follows it up with a tcp_output(). This causes the data to be
immediately sent over WiFi to GDB.