Added logging.h
Now all of the modules use the same mechanism to log errors, debug
messsages, and informational text to the user. The log messages always
list the filename, line number, and function name where the message
was generated. The debug and error messages can be disabled via
defines in the new config.h. Some other constants that we might want to
modify easily in the future have been moved here as well:
* GDB_SOCKET_PORT_NUMBER
* SWCLK_PIN
* SWDIO_PIN