Update README
Now that mri-swd is sort of functional, I have update the README to let
people know what is working (a short list) and what's not (a much longer
list). I also added a few notes about how to build the firmware, connect
GDB, and wire up to an RP2040 Pico target.