Reduce wasted SWD cycles
Not every instruction in the swd.pio program is side setting the SWCLK
pin. These show up on the logic analzyer capture as stalls in the
clocking sequence. This commit runs the PIO state machine at twice the
needed clock rate if the desired SWD clock rate is below 125MHz/4. This
means that the instructions not updating the SWCLK pin now stall for
1/2 the time they used to.

This change does require that each instruction which updates the SWCLK
pin to have a 1 cycle delay added to it. This is done in SWD::init() by
finding all side-set instructions in the machine code and adding a 1
cycle delay. This requires no direct changes to the
swd.pio program itself. It is all done on the CPU side at init time (ie.
a sort of self-modifying code approach).