Handle SWD disconnect when processing GDB commands
All of the SWD target memory read and write calls in mri_platform.cpp
are funneled through readTargetMemory() and writeTargetMemory(). This
allows for SWD_PROTOCOL errors to be detected and used to flag that
the SWD target has either been unattached or powered off. When such
errors are detected handleUnrecoverableSwdError() is called and it will:
* Empty the queue of data from GDB to MRI and inserts an empty packet.
* The previous item will cause the MRI core to start processing that
  empty packet which will end up calling Platform_HandleGDBCommand().
* Platform_HandleGDBCommand() has been modified to detect that we
  are handling an unrecoverable SWD error and return a value that will
  cause the MRI core to exit and return control to mainDebuggerLoop()
  which will keep retrying to reattach to the SWD target once it is
  reconnected or powered back up.