Fix potential recv packet truncation
Previously GDBSocket::onRecv() copied received packet data into a 2048
byte buffer and then added its bytes to the circular queue. Now it
passes the lwip pbuf directly to the CircularQueue::write() method which
walks it directly to copy the bytes into the queue. Now the received
packet will only be truncated if it somehow manages to be larger than
16k. I think before it would sometimes receive packets that were larger
than the 2k buffer and that led to lost data during FLASH loads that
caused GDB to hang during load.