Sometimes TAR is returned as 0
An assert in SWD::calculateTransferCount() was hitting an assert because
the TAR value read was lower than the starting address. I saw that
assert fire recently when the RP2040 wasn't very happy during a watchdog
reset. It turned out that no SWD errors were encountered but the TAR
read returned a zero which wasn't valid. I now catch this, log an error,
and return a transfer count of 0 so that the higher level code can
possibly retry the read/write operation.