Break on initial attach
Previously it would fail to halt the CPU on the initial attach if the
halt on attach option was enabled. This was because halting debug hadn't
yet been enabled when the halt request was made so I moved up the
g_doesHaltDebuggingNeedToBeEnabled handling in
innerDebuggerLoop().

I also modified SWD::handleTransferResponse() to handle ACK_PARITY
errors as expected further down in the function.