Added CpuCores abstraction
I added the CpuCores class to hold the code that knows how to manipulate
Cortex-M registers to implement debugging functionality, moving it out
of the MRI Platform_*() C API. This abstraction should make it easier to
support multi-core processors. A future commit will include an
additional abstraction layer, CpuCore, to represent each of the cores.