Added CpuCore abstraction
Added another level of abstraction to represent each of the cores on a
multi-core device like the RP2040. The original CpuCores class now just
figures out which of the CpuCore objects should be used for each of the
existing methods in the CpuCores class (default core, halted core, or
all of the cores).

The code is almost ready for me to actually start exposing the 2 cores
on the RP2040 as threads to GDB.