Get ARM semi-hosting API working
The MRI core already had support for the mbed LocalFileSystem which
used the ARM semi-hosting BKPT interface to communicate with the
interface chip on the mbed-LPC1768. This is why the code in MRI refers
to it as mbed semi-hosting and not ARM semi-hosting.

I pulled that mbed semi-hosting code into mri_platform.cpp since I
needed to pull the operation parameters in a different way for mri-swd
since it isn't running in the same address space as the target code.
mri-swd needs to issue SWD based memory reads to transfer over the
parameter block for each operation.

I enabled the Pico C SDK's semi-hosting driver for handling printf()
output. This means that the test pass code now uses the ARM/mbed
semi-hosting support for printf() and all of the other file I/O is
redirected to GDB via my custom newlib semi-hosting support that I
created for MRI.

I also bumped the version ahead to 0.4-20240813.0