Clear breakpoints/watchpoints on first halt
Previously the breakpoints and watchpoints were cleared in
CpuCore::initForDebugging() which means that it would clear breakpoints
that you might have set before this power up of the hardware debugger.
This could happen if the target was started with breakpoints set and
then GDB and mri-swd detached while the target was running for some
reason (like the host being power cycled). If the developer then powers
up the debugger again then it will leave the target running (usually) so
why clear the breakpoints that the user had previously set. Might lose
an important repro if mri-swd clears them like this. So this is a good
reason not to clear the breakpoint during the init path.

A clue as to a better location can be found from when GDB crashes or is
terminated while the target is running with breakpoints/watchpoints set.
If there are old breakpoints around when GDB re-attaches then it won't
know to clear them so they just sit around and their slots in the CPU's
debug hardware won't be re-used. So it would be better to clear any
existing breakpoints/watchpoints when GDB first attaches.

mri-swd now postpones clearing any active breakpoints and/or watchpoints
until the first time the core is halted after the SWD connection is lost
and reattached (and a halt has been configured for such events) or GDB
has re-attached.