Don't allow more than 1 GDB connection at a time
Previously a second GDB connection would overwrite the PCB for the first
one and cause the original GDB connection to hang on any subsequent
network traffic.

With this commit, I catch this scenario in GDBSocket::onAccept() and
just error out the second attempt so that the original one is properly
maintained.