Fix assert in RP2040's allocateTargetRam()
The following assert would fire sometimes:
  pObject->ramWrite < pObject->ramRead

It turns out this could occur because canAllocateTargetRam() was using
"<=" for the same comparison. It should use the same "<" comparison
instead.

I also bumped the version ahead from 0.4 to 0.5 since I want to make
this the LKG since it contains this fix which was sometimes hitting on
the dogfood mri-swd debugger.