Fix build issue
The custom build steps didn't know that the LocalFlashing code was
dependent on the devices/LocalFlashing.h header file and I missed the
fact that the header was pulling in C++ code which the C code to run
on the target itself couldn't compile. I added the dependency and
protected the C++ definitions with an "#ifdef __cplusplus".